backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
python-multipart==0.0.20
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
Pillow==11.1.0
pytesseract==0.3.13
watchfiles==1.0.3
pydantic==2.10.5
pydantic-settings==2.7.1
easyocr==1.7.2
google-cloud-vision==3.8.1